backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.1
mysql-connector-python==9.3.0
google-generativeai==0.8.5
requests==2.32.4
python-dotenv==1.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
